rename board files
