+++ board_test.go
@@ -1,295 +1,176 @@
 package main
 
 import (
+	"log/slog"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 )
 
-func TestArrayBoardIsCellTaken(t *testing.T) {
-	cases := []struct {
-		Input Position
-		Want  bool
-	}{
-		{Position{0, 0}, true},
-		{Position{1, 1}, false},
-		{Position{2, 2}, false},
-	}
+func TestIndexedBoard(t *testing.T) {
+	logger = NewLogger(slog.LevelInfo)
+	b := NewBoard(3)
 
-	for _, c := range cases {
-		want := c.Want
-		b := ArrayBoard{}
-
-		b.init(3)
-
-		b.FromStringCells(
-			[][]string{
-				{"n", "n", "n"},
-				{"b", "b", "b"},
-				{"w", "w", "w"},
-			},
-		)
-
-		got := b.isCellTaken(c.Input)
-
-		assert.Equal(t, want, got)
-	}
+	assert.Equal(t, 3, b.N)
 }
 
-func TestArrayBoardGetCellsToFlip(t *testing.T) {
-	b := ArrayBoard{}
+func TestIndexedBoardCalcIdxN(t *testing.T) {
+	logger = NewLogger(slog.LevelInfo)
+	b := NewBoard(3)
 
-	b.init(4)
+	idxN := b.calcLineN(8)
 
-	type input struct {
-		CellsStr [][]string
-		Position Position
-		Turn     Turn
-	}
+	assert.Equal(t, 38, idxN)
 
-	cases := []struct {
-		Name  string
-		Input input
-		Want  []CellToFlip
-	}{
-		{
-			Name: "Horizontal Right, black",
-			Input: input{
-				[][]string{
-					{"n", "w", "b", "n"},
-					{"n", "n", "n", "n"},
-					{"n", "n", "n", "n"},
-					{"n", "n", "n", "n"},
-				},
-				Position{0, 0},
-				Black,
-			},
-			Want: []CellToFlip{
-				{1, 0, HasBlack},
-			},
-		},
-		{
-			Name: "Horizontal Left, black",
-			Input: input{
-				[][]string{
-					{"b", "w", "w", "n"},
-					{"n", "n", "n", "n"},
-					{"n", "n", "n", "n"},
-					{"n", "n", "n", "n"},
-				},
-				Position{3, 0},
-				Black,
-			},
-			Want: []CellToFlip{
-				{2, 0, HasBlack},
-				{1, 0, HasBlack},
-			},
-		},
-		{
-			Name: "Vertical to bottom, black",
-			Input: input{
-				[][]string{
-					{"n", "n", "n", "n"},
-					{"w", "n", "n", "n"},
-					{"w", "n", "n", "n"},
-					{"b", "n", "n", "n"},
-				},
-				Position{0, 0},
-				Black,
-			},
-			Want: []CellToFlip{
-				{0, 1, HasBlack},
-				{0, 2, HasBlack},
-			},
-		},
-		{
-			Name: "Vertical to top, black",
-			Input: input{
-				[][]string{
-					{"n", "b", "n", "n"},
-					{"n", "w", "n", "n"},
-					{"n", "w", "n", "n"},
-					{"n", "n", "n", "n"},
-				},
-				Position{1, 3},
-				Black,
-			},
-			Want: []CellToFlip{
-				{1, 2, HasBlack},
-				{1, 1, HasBlack},
-			},
-		},
-		{
-			Name: "Diagonal to bottom right, black",
-			Input: input{
-				[][]string{
-					{"n", "n", "n", "n"},
-					{"n", "w", "n", "n"},
-					{"n", "n", "w", "n"},
-					{"n", "n", "n", "b"},
-				},
-				Position{0, 0},
-				Black,
-			},
-			Want: []CellToFlip{
-				{1, 1, HasBlack},
-				{2, 2, HasBlack},
-			},
-		},
-		{
-			Name: "Diagonal to bottom left, white",
-			Input: input{
-				[][]string{
-					{"n", "n", "n", "n"},
-					{"n", "n", "b", "b"},
-					{"n", "b", "w", "n"},
-					{"w", "n", "n", "w"},
-				},
-				Position{3, 0},
-				White,
-			},
-			Want: []CellToFlip{
-				{2, 1, HasWhite},
-				{1, 2, HasWhite},
-			},
-		},
-		{
-			Name: "Diagonal to top right, white",
-			Input: input{
-				[][]string{
-					{"n", "n", "w", "w"},
-					{"n", "b", "b", "n"},
-					{"n", "b", "n", "w"},
-					{"n", "n", "n", "n"},
-				},
-				Position{0, 2},
-				White,
-			},
-			Want: []CellToFlip{
-				{1, 1, HasWhite},
-			},
-		},
-		{
-			Name: "Diagonal to top left, black",
-			Input: input{
-				[][]string{
-					{"w", "n", "w", "w"},
-					{"n", "b", "b", "n"},
-					{"n", "b", "b", "w"},
-					{"n", "n", "n", "n"},
-				},
-				Position{3, 3},
-				White,
-			},
-			Want: []CellToFlip{
-				{2, 2, HasWhite},
-				{1, 1, HasWhite},
-			},
-		},
-		{
-			Name: "multiple directions 1",
-			Input: input{
-				[][]string{
-					{"n", "w", "w", "b"},
-					{"w", "w", "b", "w"},
-					{"w", "b", "w", "n"},
-					{"b", "n", "n", "b"},
-				},
-				Position{0, 0},
-				Black,
-			},
-			Want: []CellToFlip{
-				{1, 0, HasBlack},
-				{2, 0, HasBlack},
-				{0, 1, HasBlack},
-				{0, 2, HasBlack},
-				{1, 1, HasBlack},
-				{2, 2, HasBlack},
-			},
-		},
-		{
-			Name: "multiple directions 2",
-			Input: input{
-				[][]string{
-					{"w", "w", "w", "b"},
-					{"b", "b", "b", "w"},
-					{"n", "b", "w", "n"},
-					{"w", "w", "n", "b"},
-				},
-				Position{0, 2},
-				White,
-			},
-			Want: []CellToFlip{
-				{1, 2, HasWhite},
-				{0, 1, HasWhite},
-				{1, 1, HasWhite},
-			},
-		},
-		{
-			Name: "Do nothing horizontal",
-			Input: input{
-				[][]string{
-					{"n", "w", "w", "w"},
-					{"n", "w", "b", "n"},
-					{"n", "b", "w", "n"},
-					{"w", "n", "n", "n"},
-				},
-				Position{0, 0},
-				Black,
-			},
-			Want: []CellToFlip{},
-		},
-		{
-			Name: "Do nothing vertical",
-			Input: input{
-				[][]string{
-					{"n", "w", "b", "w"},
-					{"n", "b", "w", "w"},
-					{"w", "n", "n", "w"},
-					{"w", "w", "w", "n"},
-				},
-				Position{3, 3},
-				Black,
-			},
-			Want: []CellToFlip{},
-		},
-		{
-			Name: "Do nothing diagnol",
-			Input: input{
-				[][]string{
-					{"b", "n", "n", "n"},
-					{"n", "w", "b", "n"},
-					{"n", "b", "w", "n"},
-					{"w", "n", "n", "n"},
-				},
-				Position{0, 1},
-				White,
-			},
-			Want: []CellToFlip{},
-		},
-	}
+	idxN = b.calcLineN(3)
 
-	for _, c := range cases {
-		cellsStr, position, turn := c.Input.CellsStr, c.Input.Position, c.Input.Turn
+	assert.Equal(t, 8, idxN)
+}
 
-		b.FromStringCells(cellsStr)
-		b.Turn = turn
+func TestIndexedBoardCalcLinesForCell(t *testing.T) {
+	logger = NewLogger(slog.LevelInfo)
+	n := 4
 
-		result := b.GetCellsToFlip(position.X, position.Y)
+	idxForCells := NewLineForCells(n)
 
-		if len(result) != len(c.Want) {
-			t.Errorf("%s: Expected %d cell(s) to flip, got %d, %v", c.Name, len(c.Want), len(result), result)
-			continue
-		}
+	idxForCell5 := idxForCells[5]
 
-		for i, want := range c.Want {
-			if result[i].X != want.X ||
-				result[i].Y != want.Y ||
-				result[i].NewState != want.NewState {
-				t.Errorf("%s: Expected %v, got %v", c.Name, want, result[i])
-			}
+	assert.Equal(t, LineForCell{1, 1, Row}, idxForCell5[0])
+	assert.Equal(t, LineForCell{5, 1, Col}, idxForCell5[1])
+	assert.Equal(t, LineForCell{9, 1, BackSlash}, idxForCell5[2])
+	assert.Equal(t, LineForCell{11, 1, Slash}, idxForCell5[3])
 
-		}
+	idxForCell9 := idxForCells[9]
 
-	}
+	assert.Equal(t, LineForCell{2, 1, Row}, idxForCell9[0])
+	assert.Equal(t, LineForCell{5, 2, Col}, idxForCell9[1])
+	assert.Equal(t, LineForCell{8, 1, BackSlash}, idxForCell9[2])
+	assert.Equal(t, LineForCell{12, 2, Slash}, idxForCell9[3])
+
+	idxForCell10 := idxForCells[10]
+
+	assert.Equal(t, LineForCell{2, 2, Row}, idxForCell10[0])
+	assert.Equal(t, LineForCell{6, 2, Col}, idxForCell10[1])
+	assert.Equal(t, LineForCell{9, 2, BackSlash}, idxForCell10[2])
+	assert.Equal(t, LineForCell{13, 1, Slash}, idxForCell10[3])
+}
+
+func TestIndexedBoardCalcMobility(t *testing.T) {
+	logger = NewLogger(slog.LevelInfo)
+	n := 4
+
+	m := NewMobility(n)
+
+	assert.Equal(t, []int{0, 0}, m[Idx{0, n}][Black][0])
+	assert.Equal(t, []int{1, 0}, m[Idx{5, n}][White][2])
+	assert.Equal(t, []int{1, 0}, m[Idx{7, n}][Black][2])
+	assert.Equal(t, []int{0, 1}, m[Idx{15, n}][Black][0])
+	assert.Equal(t, []int{0, 1}, m[Idx{21, n}][White][0])
+}
+
+func TestIndexedBoardPlaceWithoutCheck(t *testing.T) {
+	logger = NewLogger(slog.LevelInfo)
+
+	n := 3
+
+	b := NewBoard(n)
+
+	// |1|2|0|
+	// |2|1|0|
+	// |0|0|0|
+	assert.Equal(t, Idx{7, n}, b.Lines[LineId(0)])
+	assert.Equal(t, Idx{5, n}, b.Lines[LineId(1)])
+	assert.Equal(t, Idx{0, n}, b.Lines[LineId(2)])
+	assert.Equal(t, Idx{7, n}, b.Lines[LineId(3)])
+	assert.Equal(t, Idx{5, n}, b.Lines[LineId(4)])
+	assert.Equal(t, Idx{0, n}, b.Lines[LineId(5)])
+	assert.Equal(t, Idx{4, n}, b.Lines[LineId(6)])
+	assert.Equal(t, Idx{3, n}, b.Lines[LineId(7)])
+
+	b.PlaceWithoutCheck(2, Black)
+
+	// |1|1|1|
+	// |2|1|0|
+	// |0|0|0|
+	assert.Equal(t, Idx{13, n}, b.Lines[LineId(0)])
+	assert.Equal(t, Idx{5, n}, b.Lines[LineId(1)])
+	assert.Equal(t, Idx{0, n}, b.Lines[LineId(2)])
+	assert.Equal(t, Idx{7, n}, b.Lines[LineId(3)])
+	assert.Equal(t, Idx{4, n}, b.Lines[LineId(4)])
+	assert.Equal(t, Idx{1, n}, b.Lines[LineId(5)])
+	assert.Equal(t, Idx{4, n}, b.Lines[LineId(6)])
+	assert.Equal(t, Idx{4, n}, b.Lines[LineId(7)])
+
+	b.PlaceWithoutCheck(5, White)
+
+	// |1|1|1|
+	// |2|2|2|
+	// |0|0|0|
+	assert.Equal(t, Idx{13, n}, b.Lines[LineId(0)])
+	assert.Equal(t, Idx{26, n}, b.Lines[LineId(1)])
+	assert.Equal(t, Idx{0, n}, b.Lines[LineId(2)])
+	assert.Equal(t, Idx{7, n}, b.Lines[LineId(3)])
+	assert.Equal(t, Idx{7, n}, b.Lines[LineId(4)])
+	assert.Equal(t, Idx{7, n}, b.Lines[LineId(5)])
+	assert.Equal(t, Idx{7, n}, b.Lines[LineId(6)])
+	assert.Equal(t, Idx{7, n}, b.Lines[LineId(7)])
+
+	b.PlaceWithoutCheck(6, Black)
+
+	// |1|1|1|
+	// |1|1|2|
+	// |1|0|0|
+	assert.Equal(t, Idx{13, n}, b.Lines[LineId(0)])
+	assert.Equal(t, Idx{22, n}, b.Lines[LineId(1)])
+	assert.Equal(t, Idx{1, n}, b.Lines[LineId(2)])
+	assert.Equal(t, Idx{13, n}, b.Lines[LineId(3)])
+	assert.Equal(t, Idx{4, n}, b.Lines[LineId(4)])
+	assert.Equal(t, Idx{7, n}, b.Lines[LineId(5)])
+	assert.Equal(t, Idx{4, n}, b.Lines[LineId(6)])
+	assert.Equal(t, Idx{13, n}, b.Lines[LineId(7)])
+
+}
+
+func TestIndexedBoardFromStringCells(t *testing.T) {
+	logger = NewLogger(slog.LevelInfo)
+	n := 3
+
+	b := NewBoard(n)
+
+	b.FromStringCells(
+		[][]string{
+			{"n", "n", "n"},
+			{"b", "b", "b"},
+			{"b", "w", "b"},
+		},
+	)
+
+	assert.Equal(t, Idx{0, n}, b.Lines[LineId(0)])
+	assert.Equal(t, Idx{13, n}, b.Lines[LineId(1)])
+	assert.Equal(t, Idx{16, n}, b.Lines[LineId(2)])
+	assert.Equal(t, Idx{12, n}, b.Lines[LineId(3)])
+	assert.Equal(t, Idx{21, n}, b.Lines[LineId(4)])
+	assert.Equal(t, Idx{12, n}, b.Lines[LineId(5)])
+	assert.Equal(t, Idx{12, n}, b.Lines[LineId(6)])
+	assert.Equal(t, Idx{12, n}, b.Lines[LineId(7)])
+}
+
+func TestIndexedBoardCount(t *testing.T) {
+	logger = NewLogger(slog.LevelInfo)
+	n := 3
+
+	b := NewBoard(n)
+
+	b.FromStringCells(
+		[][]string{
+			{"n", "n", "n"},
+			{"b", "b", "b"},
+			{"b", "w", "b"},
+		},
+	)
+
+	totalB, totalW := b.Count()
+
+	assert.Equal(t, 5, totalB)
+	assert.Equal(t, 1, totalW)
 }
